Remove brittle test and add setUp method
