+++ src/main/java/com/heatheryou/httpserver/route/handler/BuildResponse.java
@@ -1,0 +1,7 @@
+package com.heatheryou.httpserver.route.handler;
+
+import com.heatheryou.httpserver.Response;
+
+public interface IResponseBuilder {
+    Response buildResponse(int statusCode, String[] entityHeaders, String content);
+}
