+++ src/test/java/com/heatheryou/httpserver/RequestParserTest.java
@@ -1,31 +1,38 @@
 package com.heatheryou.httpserver;
 
+import org.junit.Before;
 import org.junit.Test;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.StringReader;
 
-import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.junit.Assert.*;
 
 public class RequestParserTest {
+    String requestLine;
+    StringReader stringReader;
+    BufferedReader bufferedReader;
+    RequestParser parser;
+
+    @Before
+    public void setUp() {
+        requestLine = "GET /simple_get HTTP/1.1";
+        stringReader = new StringReader(requestLine);
+        bufferedReader = new BufferedReader(stringReader);
+        parser = new RequestParser();
+    }
+
     @Test
     public void processRequestReturnsRequest() throws IOException {
-        String request = "GET /simple_get HTTP/1.1";
-        StringReader stringReader = new StringReader(request);
-        BufferedReader bufferedReader = new BufferedReader(stringReader);
-        RequestParser parser = new RequestParser();
-        Request actual = parser.processRequest(bufferedReader);
-        assertThat(actual, instanceOf(Request.class));
+        Request request = parser.processRequest(bufferedReader);
+        String[] actual = new String[]{ request.getUri(), request.getMethod() };
+        String[] expected = new String[]{"/simple_get", "GET"};
+        assertEquals(expected, actual);
     }
 
     @Test
     public void getMethodReturnsMethodOfRequest() throws IOException {
-        String request = "GET /simple_get HTTP/1.1";
-        StringReader stringReader = new StringReader(request);
-        BufferedReader bufferedReader = new BufferedReader(stringReader);
-        RequestParser parser = new RequestParser();
         parser.processRequest(bufferedReader);
         String actual = parser.getMethod();
         String expected = "GET";
@@ -34,10 +41,6 @@ public class RequestParserTest {
 
     @Test
     public void getMethodReturnsUriOfRequest() throws IOException {
-        String request = "GET /simple_get HTTP/1.1";
-        StringReader stringReader = new StringReader(request);
-        BufferedReader bufferedReader = new BufferedReader(stringReader);
-        RequestParser parser = new RequestParser();
         parser.processRequest(bufferedReader);
         String actual = parser.getUri();
         String expected = "/simple_get";
